+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All Notable changes to `flipboxdigital\pipeline` will be documented in this file
 
+## Unreleased
+### Changed
+- Processors pass along the existing `$source` string as an array of `$extra` params.
+
 ## 1.0.0
 - Initial release!